gradle-test-master/gradle-test-webapp/src/test/java/com/ts/servlet/DemoServletTest.java
package com.ts.servlet;

import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.MockitoAnnotations;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.PrintWriter;
import java.io.StringWriter;

import static org.junit.Assert.assertEquals;

/**
 * @author: Owen Jia
 * @time: 2018/11/26 13:50
 */
public class DemoServletTest {
    @Mock
    HttpServletRequest request;

    @Mock
    HttpServletResponse response;

    @Before
    public void setUp() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    public void doGet() throws Exception {
        StringWriter stringWriter = new StringWriter();
        PrintWriter printWriter = new PrintWriter(stringWriter);

        Mockito.when(response.getWriter()).thenReturn(printWriter);

        new DemoServlet().doGet(request, response);

        assertEquals("i am DemoServlet,i do test connect.", stringWriter.toString());
    }
}
